requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-decouple
psycopg2-binary
alembic
pydantic[email]